Rationalised presence/absence of form element and InputTypeSubmit for these components
